+++ src/main/java/com/ezban/Application.java
@@ -1,4 +1,4 @@
-package com;
+package com.ezban;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
