+++ src/main/java/com/ezban/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com;
+package com.ezban;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
