+++ src/main/java/com/ezban/controller/BackstageController.java
@@ -1,21 +1,20 @@
-package com.controller;
+package com.ezban.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/frontstage")
-public class FrontstageController {
+@RequestMapping("/backstage")
+public class BackstageController {
 
-    @GetMapping("/")
+    @GetMapping("")
     public String index() {
-        return "/frontstage/index2";
+        return "/backstage/index";
     }
 
     @GetMapping("/template")
     public String template() {
-        return "/frontstage/template2";
+        return "/backstage/template";
     }
-
 }
